+++ hgctl-go/internal/commands/register/register_avs.go
@@ -1,4 +1,4 @@
-package commands
+package register
 
 import (
 	"fmt"
