+++ hgctl-go/internal/commands/register/register_avs_test.go
@@ -1,4 +1,4 @@
-package commands
+package register
 
 import (
 	"testing"
