+++ hgctl-go/internal/commands/register/register_key.go
@@ -1,4 +1,4 @@
-package commands
+package register
 
 import (
 	"context"
