+++ hgctl-go/internal/commands/register/register_key_test.go
@@ -1,4 +1,4 @@
-package commands
+package register
 
 import (
 	"testing"
