+++ hgctl-go/internal/commands/register/set_allocation_delay.go
@@ -1,4 +1,4 @@
-package commands
+package register
 
 import (
 	"fmt"
