+++ hgctl-go/internal/commands/register/set_allocation_delay_test.go
@@ -1,4 +1,4 @@
-package commands
+package register
 
 import (
 	"testing"
